decoupling the tests by introducing a session parameter to the test functions
