+++ fastapi_pokedex_api/api_app.py
@@ -1,34 +1,16 @@
-import fastapi
-import requests
-from pydantic import typing
+from fastapi import FastAPI
 
+from fastapi_pokedex_api import pokedex_schema
+from fastapi_pokedex_api import pokeapi_requests
 from fastapi_pokedex_api.configs.pokeapi_config import get_pokeapi_ip
-from fastapi_pokedex_api.pokedex_schema import PokedexBase
+
+app = FastAPI()
 
 
-def get_pokedata(poke_name: str) -> PokedexBase:
-    ip = get_pokeapi_ip().POKEAPI_IP
-    name = poke_name.lower()
-    try:
-        response = requests.get("{}{}".format(ip, name))
-        response.raise_for_status()
-    except fastapi.HTTPException as httperr:
-        raise httperr
-
-    details = response.json()
-    return PokedexBase(
-        name=name,
-        description=english_description(details.get("flavor_text_entries", {})),
-        habitat=details.get("habitat").get("name")
-        if details.get("habitat") is not None
-        else None,
-        is_legendary=details.get("isLegendary"),
+@app.get(
+    "/pokedex/{poke_name}", response_model=pokedex_schema.PokedexBase, tags=["Pokemon"]
+)
+async def get_pokedata(poke_name: str) -> pokedex_schema.PokedexBase:
+    return pokeapi_requests.get_pokedata(
+        poke_name=poke_name, ip_settings=get_pokeapi_ip()
     )
-
-
-def english_description(descriptions: typing.List) -> typing.Optional[str]:
-    for entry in descriptions:
-        if entry.get("language").get("name") == "en":
-            return entry.get("flavor_text").replace("\n", " ").replace("\u000c", " ")
-    else:
-        return None
