+++ fastapi_pokedex_api/run.py
@@ -1,7 +1,7 @@
 import uvicorn
 
 from fastapi_pokedex_api.configs.pokedex_api_config import get_api_config
-from fastapi_pokedex_api.pokeapi_requests import app
+from fastapi_pokedex_api.api_app import app
 
 if __name__ == "__main__":
     api_settings = get_api_config()
