+++ pyproject.toml
@@ -17,6 +17,7 @@ pytest = "^5.2"
 mypy = "^0.931"
 black = "^22.1.0"
 pre-commit = "^2.17.0"
+requests-mock = "^1.9.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
